+++ app/components/primer/component.rb
@@ -6,5 +6,6 @@ module Primer
     include ClassNameHelper
     include FetchOrFallbackHelper
     include OcticonsHelper
+    include JoinStyleArgumentsHelper
   end
 end