+++ app/components/primer/progress_bar_component.rb
@@ -67,7 +67,8 @@ module Primer
 
         @system_arguments[:tag] = :span
         @system_arguments[:bg] = bg
-        @system_arguments[:style] = "width: #{@percentage}%;"
+        @system_arguments[:style] =
+          join_style_arguments("width: #{@percentage}%;", @system_arguments[:style])
         @system_arguments[:classes] = class_names("Progress-item", @system_arguments[:classes])
       end
     end