+++ app/components/primer/slot.rb
@@ -5,5 +5,6 @@ module Primer
   class Slot < ViewComponent::Slot
     include ClassNameHelper
     include FetchOrFallbackHelper
+    include JoinStyleArgumentsHelper
   end
 end